Properly expose getText
