+++ src/index.js
@@ -8,5 +8,6 @@ export * from './queries'
 export * from './wait'
 export * from './wait-for-element'
 export * from './matches'
+export * from './get-text'
 export * from './events'
 export * from './bind-element-to-queries'
